nvk: Clean up unused header from libdrm_nouveau
This was added in 49d6967a54fb05186b73fbedaccd0267a9bc4ab0 without a
dependency on libdrm_nouveau. If libdrm is not compiled with nouveau
enabled, the build errors here.

This is currently unused and since
2a2cfaec34f2c3f8c55b8f98116afee0519df00d removed dependency on
libdrm_nouveau, this should be gone too.

Fixes: 2a2cfaec34f ("nouveau: import libdrm_nouveau")
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/29517>