ci: fix meson install script
042ad3e09f910f4d3ebe ("ci/meson: reuse meson installation") forgot to
bump the image tags, and as a result this was merged untested and turned
out to be broken, as the fedora image.

The issue is that python in the Fedora image is not flagged as
EXTERNALLY-MANAGED, unlike what Debian does, so the
`--break-system-packages` is invalid.

Instead, remove this flag from the debian image as it makes very little
sense in a docker image.

Fixes: 042ad3e09f910f4d3ebe ("ci/meson: reuse meson installation")
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/29673>