freedreno/a6xx: Allow blit based transfers
Some apps (like firefox and chromium) seem to like to create textures
with internal format GL_RGBA8 but then upload GL_BGRA to them.  If the
app were a bit more clever it could hit the memcpy path.  But fallback
sw conversion (convert_ubytes()) is slow.  If we are going to have to
make this extra copy, do it on the gpu.

Signed-off-by: Rob Clark <robdclark@chromium.org>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/30064>