util/glsl2spirv: fixup the generated depfile when copying sources
So that the depfile contains a reference to the original source rather
than the copied one. This is necessary to avoid ninja not finding the
copy and causing spurious rebuilds when the copy has been removed, as
well as correctly tracking changes to the input files.

fixes: 988878a995cfcbca85d0af8b27d1e155c596d087

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/30132>