v3dv: free temp image created when copying with blit
This fixes a leak that happens when copying a image using blit, and the
image is a compressed one.

In this case a new image view is created that can be re-interpreted to
perform the copy, but was not properly free.

Reviewed-by: Iago Toral Quiroga <itoral@igalia.com>
Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/30161>