lima/ppir: add folding lowering for dest modifiers
Lowering to fold dest modifiers to prepare for this being disabled
in nir.
This will allow ppir to drop nir_legacy usage.
Currently we only handle clamp modifiers which was the only supported
one in ppir, but we can reuse the same for the other hardware supported
modifiers as a follow-up.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>
Reviewed-by: Vasily Khoruzhick <anarsoul@gmail.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/30332>