anv: be consistent regarding non-render engines on i915.ko
Today, on i915.ko, if Sparse Resources is disabled and the Kernel is
new enough to confirm to us that the GuC version is good, we'll expose
non-render engines, otherwise we don't.

Ever since we merged 1e47ffa373d9 ("anv/trtt: make all contexts have
the same TR-TT programming"), TR-TT is not anymore the reason why
we're not enabling non-render engines. Our performance team has
analyzed workloads and concluded enabling non-render engines is not
worth it on i915.ko today.

So here we adjust the code to do three things:
 - Stop blaming TR-TT
 - Unify the default behavior for i915.ko
 - Don't disable non-render engines when TR-TT is being used on xe.ko.

v2:
- Comments (José)

Acked-by: Felix DeGrood <felix.j.degrood@intel.com>
Reviewed-by: José Roberto de Souza <jose.souza@intel.com>
Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/30627>